foresight/requirements.txt
ipython
matplotlib
pandas
python-socketio
requests
websockets
websocket-client
pillow
pyexiftool
fluent-logger
